+++ apps/catalog/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from apps.catalog.models import Category, Product, Carousel
+from apps.catalog.models import Category, Product
 
 
 class ProductAdmin(admin.ModelAdmin):
@@ -11,10 +11,5 @@ class CategoryAdmin(admin.ModelAdmin):
     prepopulated_fields = {'slug': ('name',)}
 
 
-class CarouselAdmin(admin.ModelAdmin):
-    list_display = ('title', 'is_active')
-
-
 admin.site.register(Category, CategoryAdmin)
 admin.site.register(Product, ProductAdmin)
-admin.site.register(Carousel, CarouselAdmin)
