+++ apps/catalog/models.py
@@ -37,13 +37,3 @@ class Product (TimedModel):
 
     def __str__(self):
         return self.name
-
-
-class Carousel(TimedModel):
-    img = models.ImageField(upload_to='carousel')
-    title = models.CharField(max_length=120)
-    index = models.IntegerField()
-    is_active = models.BooleanField()
-
-    def __str__(self):
-        return self.title
