+++ apps/catalog/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, get_object_or_404
-from apps.catalog.models import Product, Category, Carousel
+from apps.catalog.models import Product, Category
 
 
 def products(request, slug=None):
@@ -10,10 +10,3 @@ def products(request, slug=None):
         products = products.filter(category=category)
     return render(request, 'catalog.html', {'products': products,
                                             'categories': categories})
-
-
-def index(request):
-    visuals = Carousel.objects.order_by('index').filter(is_active=True).all()
-    products = Product.objects.order_by('created_at').all()[:4]
-    return render(request, 'main.html', {'visuals': visuals,
-                                         'products': products})
