+++ apps/core/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import url
+from apps.core import views as core_views
+
+urlpatterns = [
+    url(r'^$', core_views.index, name='index'),
+]
