+++ root/settings.py
@@ -36,6 +36,7 @@ INSTALLED_APPS = [
     'apps.order',
     'apps.cart',
     'apps.catalog',
+    'apps.core'
 ]
 
 MIDDLEWARE = [
