+++ root/urls.py
@@ -6,7 +6,8 @@ from django.conf.urls.static import static
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^btemp/', include('apps.testview.urls')),
-    url(r'^', include('apps.catalog.urls')),
+    url(r'^catalog/', include('apps.catalog.urls')),
+    url(r'^', include('apps.core.urls')),
 ]
 
 if settings.DEBUG:
